+++ packages/components/organisms/f-checkout/test/specs/component/f-checkout-delivery.component.spec.js
@@ -1,31 +1,39 @@
 import forEach from 'mocha-each';
-import CheckoutComponent from '../../../test-utils/component-objects/f-checkout.component';
+const Checkout = require('../../../test-utils/component-objects/f-checkout.component');
+const checkout = new Checkout();
 
 describe('f-checkout "delivery" component tests', () => {
     before(() => {
-        browser.url('iframe.html?id=components-organisms--checkout-component&knob-Get%20Checkout%20Url=%2Fcheckout-delivery.json&knob-Available%20Fulfilment%20Url=%2Fcheckout-available-fulfilment.json&knob-Create%20Guest%20Url=%2Fcreate-guest.json&knob-Get%20Basket%20Url=%2Fget-basket-delivery.json&knob-Auth%20token=a&knob-Login%20Url=%2Flogin&knob-Place%20Order%20Url=%2Fplace-order.json&knob-Payment%20Page%20Url%20Prefix=%23%2Fpay&viewMode=story');
-        CheckoutComponent.waitForCheckoutComponent();
+
+        const checkoutData = {
+            type: 'delivery', 
+            isAuthenticated: true, 
+            isValid: true
+        }
+
+        checkout.open(checkoutData);
+        checkout.waitForComponent();
     });
 
     forEach(['mobileNumber', 'addressLine1', 'addressCity', 'addressPostcode'])
     .it('should display each fields error message', field => {
         // Act
-        CheckoutComponent.clearCheckoutForm(field);
-        CheckoutComponent.goToPayment();
+        checkout.clearCheckoutForm(field);
+        checkout.goToPayment();
 
         // Assert
-        expect(CheckoutComponent.isFieldErrorDisplayed(field)).toBe(true);
+        expect(checkout.isFieldErrorDisplayed(field)).toBe(true);
     });
 
     forEach(['addressLine1', 'addressLine2', 'addressCity', 'addressPostcode'])
     .it('should check if address fields exist', field => {
         // Assert
-        expect(CheckoutComponent.doesFieldExist(field)).toBe(true);
+        expect(checkout.doesFieldExist(field)).toBe(true);
     });
 
     it('should display the mandatory fields', () => {
         // Assert
-        expect(CheckoutComponent.isFieldDisplayed('mobileNumber')).toBe(true);
+        expect(checkout.isFieldDisplayed('mobileNumber')).toBe(true);
     });
 
     it('should prevent user from submitting a postcode with an illegal postcode', () => {
@@ -35,11 +43,11 @@ describe('f-checkout "delivery" component tests', () => {
         };
 
         // Act
-        CheckoutComponent.populateCheckoutForm(addressInfo);
-        CheckoutComponent.goToPayment();
+        checkout.populateCheckoutForm(addressInfo);
+        checkout.goToPayment();
 
         // Assert
-        expect(CheckoutComponent.isPostCodeTypeErrorDisplayed()).toBe(true);
+        expect(checkout.isPostcodeTypeErrorDisplayed()).toBe(true);
     });
 
     it('should enable a user to submit a postcode with correct characters', () => {
@@ -49,10 +57,10 @@ describe('f-checkout "delivery" component tests', () => {
         };
 
         // Act
-        CheckoutComponent.populateCheckoutForm(addressInfo);
-        CheckoutComponent.goToPayment();
+        checkout.populateCheckoutForm(addressInfo);
+        checkout.goToPayment();
 
         // Assert
-        expect(CheckoutComponent.isPostCodeTypeErrorDisplayed()).toBe(false);
+        expect(checkout.isPostcodeTypeErrorDisplayed()).toBe(false);
     });
 });