+++ packages/components/organisms/f-checkout/CHANGELOG.md
@@ -6,13 +6,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 v0.59.0
 ------------------------------
-*February 23, 2021*
+*February 24, 2021*
 
 ### Added
 - `placeOrder` to call the Order Placement API.
 - Redirect to the payment page after the order has been placed.
 
 ### Changed
+- Restructured component object into page object model
+- Refactored component and accessibility tests
 - Extracted svgTransform to global test utils for reference by other packages
 - How we pass the parameters to the `PATCH` Checkout request.